Refactor conftest.py
- move conftest.py to the root of the repository in order to make
  fixture accesible to tests from inside the library too.
- adds missing docstrings
- remove editor config lefover
- provide better failure message when pytest is run without the
  library being installed
- improves default_rules_collection fixture to make it use
  ansiblelint.constants.DEFAULT_RULESDIR

This is in sync with documented use from:
https://docs.pytest.org/en/stable/writing_plugins.html#conftest-py-plugins
